Feat: Show notes only for your account
Add decodeJWT(), this decodes the JWT given by the backend when you log in.
Add extra authentification in createNote & showNotebooks

+++ script/main.js
@@ -27,25 +27,29 @@ async function showNotebook(){
     titleField.value = "";
     textAreaField.value = "";
     const notebook = await fetchNotebooks();
+    const user = decodeJWT();
+    const account_id = user.account_id;
 
     notebook.forEach(note => {
-        const limitedDescription = note.description.length > 100
-        ? note.description.slice(0,30) + "..."
-        : note.description;
-
-        tBody.innerHTML += 
-        `
-        <tr class="notebook-single-tr">
-            <div class="notebook-single-element" data-id="${note.id}">
-                <td class="title-container">${note.title}</td>
-                <td class="description-container">${note.description}</td>
-                <td class="actions">
-                    <i onClick="fetchNote(this)" class="fas fa-edit"></i>
-                    <i onClick="deleteNote(this)" class="fas fa-trash-alt"></i>
-                </td>
-            </div>
-        </tr>
-        `
+        if(note.account_id === account_id){
+            const limitedDescription = note.description.length > 100
+            ? note.description.slice(0,30) + "..."
+            : note.description;
+    
+            tBody.innerHTML += 
+            `
+            <tr class="notebook-single-tr">
+                <div class="notebook-single-element" data-id="${note.id}">
+                    <td class="title-container">${note.title}</td>
+                    <td class="description-container">${note.description}</td>
+                    <td class="actions">
+                        <i onClick="fetchNote(this)" class="fas fa-edit"></i>
+                        <i onClick="deleteNote(this)" class="fas fa-trash-alt"></i>
+                    </td>
+                </div>
+            </tr>
+            `
+        }
 
     })};
 
@@ -126,6 +130,8 @@ async function createNote(){
     dataMode = false;
     let titleField = document.getElementById("title-field").value;
     let textAreaField = document.getElementById("text-area-field").value;
+    const user = decodeJWT();
+    const account_id = user.account_id;
     
     try{
         const response = await fetch(`${APIUrl}/submit` , {
@@ -133,15 +139,25 @@ async function createNote(){
         headers: {
             'Content-Type': 'application/json',
         },
-        body: JSON.stringify({title: titleField, description: textAreaField})
+        body: JSON.stringify({title: titleField, description: textAreaField, account_id:account_id})
     })}catch(error){
             alert("There was an error. Please try again.")
             console.error(error);
     }
-    const result = await response.json();
     textAreaField.value = "";
     titleField.value = "";
     location.reload();
+}
+
+function decodeJWT(){
+    const token = sessionStorage.getItem('validationToken');
+    if(!token){
+        return;
+    }
+
+    const base64URL = token.split(".")[1];
+    const base64 = base64URL.replace("-", "+").replace("_", "/");
+    return JSON.parse(window.atob(base64));
 }
 
 document.addEventListener('DOMContentLoaded', () =>{
